backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

